:recycle: Chore: Productionize vm package
