+++ .changeset/forty-planes-walk.md
@@ -1,0 +1,5 @@
+---
+"@tevm/vm": patch
+---
+
+Exported more utils and actions from vm package
