+++ packages/actions/src/Call/buildAndCachePendingBlock.js
@@ -1,0 +1,10 @@
+/**
+* @param {import("@tevm/base-client").BaseClient} client
+*/
+export const buildAndCachePendingBlock = async (client) => {
+const vm = await client.getVm().then(vm => vm.deepCopy())
+const txPool = await client.getTxPool()
+txPool
+let iterations = 0
+while ()
+}
