+++ packages/vm/src/actions/BuildStatus.ts
@@ -1,0 +1,5 @@
+export enum BuildStatus {
+	Reverted = 'reverted',
+	Build = 'build',
+	Pending = 'pending',
+}
