+++ packages/vm/src/actions/runBlock.ts
@@ -32,7 +32,7 @@ import type {
 	RunBlockResult,
 	RunTxResult,
 	TxReceipt,
-} from '../utils/types.js'
+} from '../utils/index.js'
 import { runTx } from './runTx.js'
 
 const parentBeaconBlockRootAddress = EthjsAddress.fromString('0x000F3df6D732807Ef1319fB7B8bB8522d0Beac02')
