+++ packages/vm/src/utils/AfterblockEvent.ts
@@ -1,0 +1,7 @@
+import type { Block } from '@tevm/block'
+import type { RunBlockResult } from './RunBlockResult.js'
+
+export interface AfterBlockEvent extends RunBlockResult {
+	// The block which just finished processing
+	block: Block
+}
