+++ packages/vm/src/utils/ApplyBlockResult.ts
@@ -1,0 +1,34 @@
+import type { Bloom } from '@ethereumjs/vm'
+import type { Hex } from '@tevm/utils'
+import type { TxReceipt } from './TxReceipt.js'
+import type { RunTxResult } from './RunTxResult.js'
+
+/**
+ * Result of {@link applyBlock}
+ */
+export interface ApplyBlockResult {
+	/**
+	 * The Bloom filter
+	 */
+	bloom: Bloom
+	/**
+	 * The gas used after executing the block
+	 */
+	gasUsed: bigint
+	/**
+	 * The receipt root after executing the block
+	 */
+	receiptsRoot: Uint8Array
+	/**
+	 * Receipts generated for transactions in the block
+	 */
+	receipts: TxReceipt[]
+	/**
+	 * Results of executing the transactions in the block
+	 */
+	results: RunTxResult[]
+	/**
+	 * Preimages mapping of the touched accounts from the block (see reportPreimages option)
+	 */
+	preimages?: Map<Hex, Uint8Array>
+}
