+++ packages/vm/src/utils/EVMProfileOpts.ts
@@ -1,0 +1,4 @@
+export type EVMProfilerOpts = {
+	enabled: boolean
+	// extra options here (such as use X hardfork for gas)
+}
