+++ packages/vm/src/utils/SealBlockOpts.ts
@@ -1,0 +1,16 @@
+/**
+ * Options for sealing a block.
+ */
+export interface SealBlockOpts {
+	/**
+	 * For PoW, the nonce.
+	 * Overrides the value passed in the constructor.
+	 */
+	nonce?: Uint8Array
+
+	/**
+	 * For PoW, the mixHash.
+	 * Overrides the value passed in the constructor.
+	 */
+	mixHash?: Uint8Array
+}
