+++ packages/vm/src/utils/TxReceipt.ts
@@ -1,0 +1,5 @@
+import type { PostByzantiumTxReceipt } from './PostByzantiumTxReceipt.js'
+import type { PreByzantiumTxReceipt } from './PrebyzantiumTxReceipt.js'
+import type { EIP4844BlobTxReceipt } from './EIP4844BlobTxReceipt.js'
+
+export type TxReceipt = PreByzantiumTxReceipt | PostByzantiumTxReceipt | EIP4844BlobTxReceipt
