+++ packages/vm/src/utils/VMProfileOpts.ts
@@ -1,0 +1,5 @@
+export type VMProfilerOpts = {
+	//evmProfilerOpts: EVMProfilerOpts
+	reportAfterTx?: boolean
+	reportAfterBlock?: boolean
+}
